requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
lxml>=4.9.0
urllib3>=2.0.0
python-dotenv>=1.0.0
tls-client>=0.2.0
cloudscraper>=1.2.71
playwright>=1.40.0
playwright-stealth>=1.0.0
pydantic>=2.0.0
markdownify>=0.11.0
markdown2>=2.4.0
aiohttp>=3.9.0

# Job Matcher dependencies
pyyaml>=6.0
jinja2>=3.1.2
python-docx>=1.1.0
PyPDF2>=3.0.0

# Email
google_auth_oauthlib
google-api-python-client

# CLI dependencies
click>=8.1.0
rich>=13.0.0
questionary>=2.0.0

# MCP Server Dependencies

# Core dependencies (if not already installed)
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# HTTP client for testing
httpx>=0.25.0

# YAML support (usually already installed)
pyyaml>=6.0

# Web UI dependencies
flask>=3.0.0
flask-cors>=4.0.0
markdown>=3.5.0
